Extensible

Description

Generates an extensible datatype from a datatype declaration, roughly following the pattern given by the Trees that Grow paper by Najd and Peyton Jones.

A type family is generated for each constructor, taking an argument named ext for the extension type, followed by the arguments of the datatype. The names of the type families correspond to the constructors themselves modified with annotationName.

An extra type family is generated with the same arguments, named after the datatype modified with extensionName.

The datatype itself is renamed according to datatypeName and given an extra argument called ext (before the others).

Each existing constructor is renamed according to constructorName, and given an extra strict field of the corresponding type family generated above.

An extra constructor is generated for the extension type family (with the same name), containing it as its sole field.

Due to GHC's staging restriction, it is not possible to write extensible [d| data Foo = ... |] and use the generated extendFoo function within the same module.

The module where extensible is called needs the following extensions to be enabled:

TemplateHaskell,

TypeFamilies,

ConstraintKinds (TODO), and

KindSignatures (TODO).

Modules calling extendFoo need:

TemplateHaskell,

TypeFamilies,

PatternSynonyms, and

StandaloneDeriving (TODO).

You will probably also currently want to disable the warning for missing pattern type signatures (-Wno-missing-pattern-synonym-signatures).

Example

module Foo.Base where 
import Extensible

extensible [d| data Foo a = Bar a | Baz (Foo a) (Foo Int) |]

====>

type family XBar ext a 
type family XBaz ext a 
type family FooX ext a 

data Foo' ext a = 
    Bar' a !(XBar ext a) 
  | Baz' (Foo' ext a) (Foo' ext Int) !(XBaz ext a) 
  | FooX !(FooX ext a) 

data ExtFoo = ExtFoo { 
    nameBar  :: String,                        
    typeBar  :: Maybe (TypeQ -> TypeQ),    
    nameBaz  :: String,                        
    typeBaz  :: Maybe (TypeQ -> TypeQ),    
    typeFooX :: [(String, TypeQ -> TypeQ)] 
  }

defaultExtFoo :: ExtFoo 
defaultExtFoo = ExtFoo {
    nameBar  = "Bar",
    typeBar  = Just $ \_ -> [t| () |],
    nameBaz  = "Baz",
    typeBaz  = Just $ \_ -> [t| () |],
    typeFooX = []
  }

extendFoo :: String -- ^ Type alias name  
          -> TypeQ  -- ^ Tag for this annotation
          -> ExtFoo
          -> DecsQ
extendFoo name tag exts = ...

module Foo (module Foo.Base, module Foo) where
import Foo.Base

data QZ 

extendFoo "Foo" [t|QZ|] $ defaultExtFoo {
  typeBar = Nothing,  -- disable Bar
  typeFooX =          -- add two new constructors, Quux and Zoop
    [("Quux", \_ -> [t|Int|]),
     ("Zoop", \a -> [t|Foo' QZ $a|])]
}

====>

type instance XBar QZ a = Void
type instance XBaz QZ a = ()
type instance FooX QZ a = Either Int Bool

type Foo = Foo' QZ 

-- no pattern for Bar'

pattern Baz :: Foo a -> Foo Int -> Foo a 
pattern Baz x y = Baz' x y ()

pattern Quux :: Int -> Foo a 
pattern Quux x = FooX (Left x)

pattern Zoop :: Foo a -> Foo a 
pattern Zoop x = FooX (Right x)

{-# COMPLETE Baz, Quux, Zoop #-}

Name manipulation

data NameAffix

Extra strings to add to the beginning and/or end of (the base part of) Names

Constructors

NameAffix

Fields

naPrefix, naSuffix :: String

Bundled Patterns

pattern NamePrefix :: String -> NameAffix
Just a prefix, with an empty suffix

pattern NameSuffix :: String -> NameAffix
Just a suffix, with an empty prefix

applyAffix :: NameAffix -> Name -> Name

>>> applyAffix (NameAffix "pre" "Suf") (mkName "Foo")
preFooSuf
>>> applyAffix (NameAffix "pre" "Suf") (mkName "Foo.Bar")
Foo.preBarSuf

Generating extensible datatypes

extensible :: DecsQ -> DecsQ

As extensibleWith, using defaultConfig.

extensibleWith :: Config -> DecsQ -> DecsQ

Generate an extensible datatype using the given Config for creating names. See the module documentation for more detail on what this function spits out.

data Config

Configuration options for how to name the generated constructors, type families, etc.

Constructors

Config

Fields

datatypeName :: NameAffix

Applied to input datatype's name to get extensible type's name

constructorName :: NameAffix

Appled to input constructor names to get extensible constructor names

annotationName :: NameAffix

Appled to constructor names to get the annotation type family's name

extensionName :: NameAffix

Applied to datatype name to get extension constructor & type family's name

extRecordName :: NameAffix

Applied to datatype name to get extension record name

extRecTypeName :: NameAffix

Applied to constructor names to get the names of the type fields in the extension record

extRecNameName :: NameAffix

Applied to constructor names to get the names of the name fields in the extension record (which are used to name the pattern synonyms)

defExtRecName :: NameAffix

Applied to the extRecordName to get the name of the default extension

extFunName :: NameAffix

Applied to datatype name to get the name of the extension generator function

defaultConfig :: Config

Default config:

Config {
  datatypeName    = NameSuffix "'",
  constructorName = NameSuffix "'",
  annotationName  = NamePrefix "X",
  extensionName   = NameSuffix "X",
  extRecordName   = NamePrefix "Ext",
  extRecTypeName  = NamePrefix "type",
  extRecNameName  = NamePrefix "name",
  defExtRecName   = NamePrefix "default",
  extFunName      = NamePrefix "extend"
}

data ConAnn t

An annotation for a constructor. t is TypeQ -> ... -> TypeQ with one argument for each type variable in the original datatype declaration.

Ann: the annotation is the given type

NoAnn: no annotation (filled in with () automatically by the pattern synonym)

Disabled: constructor disabled (annotation type is Void and no pattern synonym generated)

Constructors

Ann t

NoAnn

Disabled

Orphan instances

Lift Name

Instance details

Methods

lift :: Name -> Q Exp

Lift ModName

Instance details

Methods

lift :: ModName -> Q Exp

Lift PkgName

Instance details

Methods

lift :: PkgName -> Q Exp

Lift OccName

Instance details

Methods

lift :: OccName -> Q Exp

Lift NameFlavour

Instance details

Methods

lift :: NameFlavour -> Q Exp

Lift NameSpace

Instance details

Methods

lift :: NameSpace -> Q Exp